Feat: 주점업자측 체크아웃 컨펌창 구현

+++ src/components/store/ItemParticipant.tsx
@@ -1,7 +1,13 @@
 import styled from 'styled-components';
+import Swal from 'sweetalert2';
+import withReactContent from 'sweetalert2-react-content';
 import thema from '../../styles/thema';
 import { deleteParticipantByStore } from '../../utils/timeDeal';
+import useSocket from '../../utils/useSocket';
 import useTimeDealStore from '../../stores/store/timeDealStore';
+import ConfirmPopup from './popup/ConfirmPopup';
+import SuccessPopup from './popup/SuccessPopup';
+import FailPopup from './popup/FailPopup';
 import type { MiniUserDTO } from '../../utils/interface';
 
 const Container = styled.div`
@@ -43,16 +49,63 @@ function ItemParticipant({
   timeDealId: number;
 }) {
   const { removeParticipant } = useTimeDealStore();
+  const MySwal = withReactContent(Swal);
+  const token = localStorage.getItem('token') as string;
+  const { socket } = useSocket(token);
 
   async function checkOut() {
-    const response = await deleteParticipantByStore(item.id);
-    if (typeof response === 'boolean') {
-      // TODO: 타임딜 아이템의 participant목록에서 삭제하기
-      removeParticipant(timeDealId, item.id);
-    } else {
-      // TODO: 팝업창 띄우기
-      console.log(response.message);
-    }
+    MySwal.fire({
+      html: (
+        <ConfirmPopup
+          title="체크아웃"
+          description={'체크아웃하시겠습니까?'}
+          confirm={Swal.clickConfirm}
+          close={Swal.close}
+        />
+      ),
+      showConfirmButton: false,
+      width: '480px',
+      padding: 0,
+      customClass: {
+        popup: 'fail-popup-border',
+      },
+    }).then(async (result) => {
+      if (result.isConfirmed) {
+        const response = await deleteParticipantByStore(item.id);
+        if (typeof response === 'boolean') {
+          removeParticipant(timeDealId, item.id);
+          MySwal.fire({
+            html: (
+              <SuccessPopup
+                title="체크아웃"
+                description="체크아웃에 성공했습니다."
+                close={Swal.clickCancel}
+              />
+            ),
+            showConfirmButton: false,
+            width: '480px',
+            padding: 0,
+            customClass: {
+              popup: 'fail-popup-border',
+            },
+            timer: 2000,
+          });
+        } else {
+          MySwal.fire({
+            html: (
+              <FailPopup title="체크아웃" description={response.message} close={Swal.clickCancel} />
+            ),
+            showConfirmButton: false,
+            width: '480px',
+            padding: 0,
+            customClass: {
+              popup: 'fail-popup-border',
+            },
+            timer: 2000,
+          });
+        }
+      }
+    });
   }
 
   return (
